Support building VSCode extension on windows
Resolves SassError: Can't find stylesheet to import.
  ╷
1 │ @use "E:\Documents\repositories\CodeWebChat-Official\packages\ui\src\styles\foundation" as *;
  │ 

Per Gemini:
Absolute Path: The build process is trying to import a SASS file using an absolute Windows path (E:\...). This is a major red flag in a portable project. Imports should almost always be relative or use configured aliases.

SASS and Backslashes: SASS, like many tools originating in the Unix world, can be particular about path separators. When it sees a quoted string with backslashes (\), it can misinterpret them as escape sequences instead of path separators. For example, \f could be interpreted as a form feed character. This mangles the path, and SASS is unable to find the file, resulting in the Can't find stylesheet to import error.